Version 6
Login

+++ Laravel/ProyectoLogUCAB/app/Http/Controllers/ClientController.php
@@ -3,9 +3,13 @@
 namespace LogUCAB\Http\Controllers;
 
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Auth;
 use LogUCAB\Http\Requests;
 use LogUCAB\Client;
+use LogUCAB\User;
+use LogUCAB\Usuario;
+use LogUCAB\Rol;
+use LogUCAB\Priv_Rol;
 use Illuminate\Support\Facades\Input;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Foundation\Validation\ValidatesRequests;
@@ -15,7 +19,7 @@ use DB;
 
 class ClientController extends Controller
 {
-    public function __construct(){
+    public function __construct(){$this->middleware('auth');
     }
 
     public function inicio(){
@@ -23,6 +27,10 @@ class ClientController extends Controller
     }
 
     public function create(){
+        if(null !== Auth::user()){
+            $Correo_Personal = Auth::user()->email;
+            return view("persona.cliente.createclient", compact('Correo_Personal'));
+        }
         return view("persona.cliente.createclient");
     }
 
@@ -34,6 +42,9 @@ class ClientController extends Controller
             $request->FK_Asignado_Tipo = 3;
         }
         $cliente = Client::find($request->Cedula);
+        if($cliente->FK_Asignado_Tipo == 2 && $request->FK_Asignado_Tipo != 1){
+            $request->FK_Asignado_Tipo = 2;
+        }
 
         if(isset($cliente)){
             Session::flash('message','Ya existe un cliente con la cédula: "'.$request->Cedula.'".');
@@ -46,7 +57,7 @@ class ClientController extends Controller
                 return Redirect::back()->withInput(Input::all());
             }
         }
-
+        
         Client::create([
         'Cedula' => $request->Cedula,
         'Nombre' => $request->Nombre,
@@ -59,9 +70,17 @@ class ClientController extends Controller
         'Frecuente' => $request->Frecuente,
         'FK_Asignado_Tipo' => $request->FK_Asignado_Tipo
         ]);
+        $user = User::where('email', $request->Correo_Personal)->first();
+        if(isset($user)){
+            $user->rol = $request->FK_Asignado_Tipo;
+            $user->save();
+            $user = Usuario::where('Correo_Personal', $request->Correo_Personal)->first();
+            $user->rol = $request->FK_Asignado_Tipo;
+            $user->save();
+        }
 
         Session::flash('message','Cliente creado correctamente.');
-        return Redirect::to('/cliente/lista');
+        return Redirect::to('/');
     }
     
     public function lista(){
@@ -76,12 +95,20 @@ class ClientController extends Controller
     }
 
     public function actualizar(Request $request){
+        $priv = Priv_Rol::where('FK_Accede_Sis',Auth::user()->rol)
+        ->where('FK_Opcion',12)
+        ->first();
+
+    if(isset($priv)){
         $clienteOG = Client::find($request->CedulaAnt);
         $cliente = Client::find($request->Cedula);
         if($request->L_Vip){
             $request->FK_Asignado_Tipo = 1;
         }else{
             $request->FK_Asignado_Tipo = 3;
+        }
+        if($clienteOG->FK_Asignado_Tipo == 2 && $request->FK_Asignado_Tipo != 1){
+            $request->FK_Asignado_Tipo = 2;
         }
         
         if(isset($cliente) && $cliente != $clienteOG){
@@ -106,14 +133,35 @@ class ClientController extends Controller
         $clienteOG->Frecuente = $request->Frecuente;
         $clienteOG->FK_Asignado_Tipo = $request->FK_Asignado_Tipo;
         $clienteOG->save();
+        $user = User::where('email', $request->Correo_Personal)->first();
+        if(isset($user)){
+            $user->rol = $request->FK_Asignado_Tipo;
+            $user->save();
+            $user = Usuario::where('Correo_Personal', $request->Correo_Personal)->first();
+            $user->rol = $request->FK_Asignado_Tipo;
+            $user->save();
+        }
 
         Session::flash('message','Cliente modificado correctamente.');
         return Redirect::to('/cliente/lista');
+    }else{
+        Session::flash('message','Usted no tiene permisos para realizar esta accion.');
+        return Redirect::back()->withInput(Input::all());
+    }
     }
 
     public function delete($Codigo){
+        $priv = Priv_Rol::where('FK_Accede_Sis',Auth::user()->rol)
+        ->where('FK_Opcion',8)
+        ->first();
+
+    if(isset($priv)){
         DB::table('Cliente')->where('Cedula', $Codigo)->delete();
         Session::flash('messagedel','Cliente eliminado correctamente.');
         return redirect('/cliente/lista');
+    }else{
+        Session::flash('message','Usted no tiene permisos para realizar esta accion.');
+        return Redirect::back()->withInput(Input::all());
+    }
     }
 }
